10 – Blueprint Multiplayer: Find a Match Design
Before we jump into the Find a Match screen layout, we fix an issue with
the Options Menu so that our settings persist when the game is closed
and re-loaded. With that out of the way, we turn our attention to the
visual layout of our Find a Match screen which allows us to toggle
whether or not we are searching for LAN or Internet games.